Restructure directory
Renamed base 'components' folder in ./src/ to 'projects' and updated babel.config aliases.
