+++ App.tsx
@@ -1,15 +1,11 @@
 import { StyleSheet, View } from "react-native";
 import { StatusBar } from "expo-status-bar";
-import { Reactive1, Reactive2, Reactive3, Reactive7, Trimmer } from "~components";
+import { Trimmer } from "~projects";
 
 export default function App() {
   return (
     <View style={styles.container}>
       <StatusBar hidden={true} />
-      {/* <Reactive1 /> */}
-      {/* <Reactive2 /> */}
-      {/* <Reactive3 /> */}
-      {/* <Reactive7 /> */}
       <Trimmer />
     </View>
   );
