+++ src/projects/Trimmer/Grip.tsx
