+++ src/projects/Trimmer/GripIcon.tsx
