+++ src/projects/Trimmer/Trimmer.tsx
