+++ src/projects/Trimmer/Types.ts
