+++ src/projects/Trimmer/index.tsx
